pyproject.toml
[project]
name = "capstoneproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Cristianbergamo",email = "bergamo.1908@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djoser (>=2.3.1,<3.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
